Portfolio-Repository2/resume_builder/requirements.txt
django-ratelimit==4.1.0
Markdown==3.7
openai==1.65.4
pdfkit==1.0.0
python-decouple
